+++ string/ft_striteri.c
@@ -6,7 +6,7 @@
 /*   By: chlimous <chlimous@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/28 21:00:21 by chlimous          #+#    #+#             */
-/*   Updated: 2023/10/28 21:10:44 by chlimous         ###   ########.fr       */
+/*   Updated: 2024/01/05 23:46:38 by chlimous         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char *))
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
